Доделал функцию автоматического шага.
